Added some documentation
Added Initialization code for Vulkan

Removed VulkanRenderingBackend.h and renamed it's .cpp file to RenderingBackend.cpp because there isnt a need for the files to be seperate.

Moved the header code of the VulkanRenderingBackend to the RenderingBackend.h file to reduce file count and follow code conventions.

Updated the CMakeLists.txt to match